data structures,greedy,*1100

+++ Codeforces Round 805 (Div. 3)/C. Train and Queries.cpp
@@ -69,7 +69,7 @@ void solve(void){
     
     ll n; cin >> n;
 
-    cout << (n % 2 ? 0 : pow(2, n / 2));
+    cout << (n % 2 ? 0 : (ll)pow(2, n / 2));
 
     nl;
 }
